+++ nrf_802154/common/include/nrf_802154_config.h
@@ -35,12 +35,13 @@
 #ifndef NRF_802154_CONFIG_H__
 #define NRF_802154_CONFIG_H__
 
+/* This file must not include nrf.h or nrfx.h directly or indirectly,
+ * to avoid circular dependencies.
+ */
+
 #ifdef NRF_802154_PROJECT_CONFIG
 #include NRF_802154_PROJECT_CONFIG
 #endif
-
-#include <nrfx.h>
-#include "nrf_802154_nrfx_addons.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -497,19 +498,6 @@ extern "C" {
  */
 #ifndef NRF_802154_ENCRYPTION_ENABLED
 #define NRF_802154_ENCRYPTION_ENABLED 1
-#endif
-
-/**
- * @def NRF_802154_ENCRYPTION_ACCELERATOR_ECB
- *
- * Enables ECB peripheral to be used as hardware accelerator for on-the-fly AES-CCM* encryption.
- */
-#if !defined(NRF_802154_ENCRYPTION_ACCELERATOR_ECB) || defined(__DOXYGEN__)
-#if defined(NRF52_SERIES) || defined(NRF5340_XXAA) || defined(__DOXYGEN__)
-#define NRF_802154_ENCRYPTION_ACCELERATOR_ECB 1
-#elif defined(NRF54H_SERIES) || defined(NRF54L_SERIES)
-#define NRF_802154_ENCRYPTION_ACCELERATOR_ECB 0
-#endif
 #endif
 
 /**